+++ packages/security/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.3.0] (2026-04-04)
+
+### Features
+
+- **worker-run:** tighten `precomputedGap` schema to structural `GapAnalysisResult` validation ([c0361b1](https://github.com/orestes-garcia-martinez/clawos/commit/c0361b11742454861284fb976aa1fe794d82a09b))
+
 ## [0.2.2] (2026-03-31)
 
 ### Bug Fixes